+++ src/app/style/app.detail.css
@@ -1,0 +1,19 @@
+.wrapper {
+    display: flex;
+    justify-content: space-around;
+}
+
+.currentPrice {
+    display: flex;
+    justify-content: space-between;
+}
+
+img {
+    width: 60px;
+    height: 60px;
+}
+/* 
+div {
+    justify-self: center;
+    margin: auto 0;
+} */
