+++ src/app/style/app.home.css
@@ -20,7 +20,7 @@ li a {
     font-family: sans-serif;
 }
 
-arrow {
+div {
     justify-self: center;
     margin: auto 0;
 }
